Add files via upload
This personal portfolio website showcases Pallem Parusharamudu's skills and projects in a modern and visually appealing layout. Designed with basic HTML and CSS, the portfolio includes a dynamic header featuring a profile photo and an introduction, highlighting Pallem's background as a BTech student in Information Technology at VJIT College, Hyderabad. The portfolio section presents selected projects with enhanced styling and spacing, using vibrant colors and rounded elements to create a clean and professional look. The contact section is designed with user interaction in mind, offering a form for inquiries and displaying a sample email for direct communication. The site’s footer provides a polished finish with consistent design elements. With a focus on clarity and usability, this portfolio effectively demonstrates Pallem's work and contact details, while maintaining a cohesive and modern design.